# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inner Circle Partners Portal API",
    "description": "API documentation for the Inner Circle Partners Portal - a comprehensive partner referral management system.\n\n## Authentication\n\nMost endpoints require authentication via Clerk. Include the session token in your requests.\n\n## Rate Limiting\n\nAPI requests are rate-limited to prevent abuse:\n- General endpoints: 100 requests per minute\n- Auth endpoints: 10 requests per minute\n- Webhook endpoints: 1000 requests per minute\n\n## Response Format\n\nAll responses follow a standard format:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"meta\": { \"timestamp\": \"...\" }\n}\n```\n\nError responses:\n```json\n{\n  \"success\": false,\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human readable message\"\n  }\n}\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Inner Circle Support",
      "email": "support@innercircle.co"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://innercircle.co/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.innercircle.co",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/payments/payout": {
      "post": {
        "tags": ["Payments"],
        "summary": "Request payout",
        "description": "Request a payout of available earnings",
        "operationId": "requestPayout",
        "security": [{ "clerkAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["partnerId", "amount"],
                "properties": {
                  "partnerId": { "type": "string" },
                  "amount": { 
                    "type": "integer",
                    "description": "Amount in cents",
                    "minimum": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payout requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
    "/api/payments/webhook": {
      "post": {
        "tags": ["Payments"],
        "summary": "Stripe webhook",
        "description": "Receives Stripe webhook events",
        "operationId": "stripeWebhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook processed" },
          "400": { "description": "Invalid webhook" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      },
      "adminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Requires admin role in Clerk user metadata"
      }
    },
    "schemas": {
...
      "Payout": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "amount": { "type": "integer", "description": "Amount in cents" },
          "fee": { "type": "integer" },
          "netAmount": { "type": "integer" },
          "status": { "type": "string", "enum": ["PENDING", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED"] },
          "requestedAt": { "type": "string", "format": "date-time" },
          "processedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PayoutResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "payout": { "$ref": "#/components/schemas/Payout" }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            }
          }
        }
      }
    },
    "responses": {
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid request data",
                "details": {
                  "errors": {
                    "email": ["Invalid email format"]
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}